test: Run mypy static code analysis
Refactoring some of the code after running mypy static code analysis
