+++ staslib/service.py
@@ -12,15 +12,137 @@ which the Staf and the Stac objects are derived.'''
 import json
 import logging
 import pathlib
-import systemd.daemon
+from itertools import filterfalse
 import dasbus.error
 import dasbus.client.observer
 import dasbus.client.proxy
 
 from gi.repository import GLib
 from libnvme import nvme
+from systemd.daemon import notify as sd_notify
 from staslib import avahi, conf, ctrl, defs, gutil, iputil, stas, trid, udev
 
+# ******************************************************************************
+class CtrlTerminator:
+    '''The Controller Terminator is used to gracefully disconnect from
+    controllers. All communications with controllers is handled by the kernel.
+    Once we make a request to the kernel to perform an operation (e.g. connect),
+    we have to wait for it to complete before requesting another operation. This
+    is particularly important when we want to disconnect from a controller while
+    there are pending operations, especially a pending connect.
+
+    The "connect" operation is especially unpredictable because all connect
+    requests are made through the blocking interface "/dev/nvme-fabrics". This
+    means that once a "connect" operation has been submitted, and depending on
+    how many connect requests are made concurrently, it can take several seconds
+    for a connect to be processed by the kernel.
+
+    While connect or other operations are being performed, it is possible
+    that a disconnect may be requested (e.g. someone or something changes the
+    configuration to remove a controller). Because it is not possible to
+    terminate a pending operation request, we have to wait for it to complete
+    before we can issue a disconnect. Failure to do that will result in
+    operations being performed by the kernel in reverse order. For example,
+    a disconnect may be executed before a pending connect has had a chance to
+    complete. And this will result in controllers that are supposed to be
+    disconnected to be connected without nvme-stas knowing about it.
+
+    The Controller Terminator is used when we need to disconnect from a
+    controller. It will make sure that there are no pending operations before
+    issuing a disconnect.
+    '''
+
+    DISPOSAL_AUDIT_PERIOD_SEC = 30
+
+    def __init__(self):
+        self._udev = udev.UDEV
+        self._controllers = list()  # The list of controllers to dispose of.
+        self._audit_tmr = gutil.GTimer(self.DISPOSAL_AUDIT_PERIOD_SEC, self._on_disposal_check)
+
+    def dispose(self, controller: ctrl.Controller, on_controller_removed_cb, keep_connection: bool):
+        '''Invoked by a service (stafd or stacd) to dispose of a controller'''
+        if controller.all_ops_completed():
+            logging.debug(
+                'CtrlTerminator.dispose()           - %s | %s: Invoke disconnect()', controller.tid, controller.device
+            )
+            controller.disconnect(on_controller_removed_cb, keep_connection)
+        else:
+            logging.debug(
+                'CtrlTerminator.dispose()           - %s | %s: Add controller to garbage disposal',
+                controller.tid,
+                controller.device,
+            )
+            self._controllers.append((controller, keep_connection, on_controller_removed_cb, controller.tid))
+
+            self._udev.register_for_action_events('add', self._on_kernel_events)
+            self._udev.register_for_action_events('remove', self._on_kernel_events)
+
+            if self._audit_tmr.time_remaining() == 0:
+                self._audit_tmr.start()
+
+    def info(self):
+        '''@brief Get info about this object (used for debug)'''
+        info = {
+            'controllers': str([str(tid) for _, _, _, tid in self._controllers]),
+            'audit timer': str(self._audit_tmr),
+        }
+        return info
+
+    def kill(self):
+        '''Stop Controller Terminator and release resources.'''
+        self._audit_tmr.stop()
+        self._audit_tmr = None
+
+        if self._udev:
+            self._udev.unregister_for_action_events('add', self._on_kernel_events)
+            self._udev.unregister_for_action_events('remove', self._on_kernel_events)
+            self._udev = None
+
+        for controller, keep_connection, on_controller_removed_cb, _ in self._controllers:
+            controller.disconnect(on_controller_removed_cb, keep_connection)
+
+        self._controllers.clear()
+
+    def _on_kernel_events(self, udev_obj):  # pylint: disable=unused-argument
+        logging.debug('CtrlTerminator._on_kernel_events() - %s event received', udev_obj.action)
+        self._disposal_check()
+
+    def _on_disposal_check(self, *_user_data):
+        logging.debug('CtrlTerminator._on_disposal_check()- Periodic audit')
+        return GLib.SOURCE_REMOVE if self._disposal_check() else GLib.SOURCE_CONTINUE
+
+    @staticmethod
+    def _keep_or_terminate(args):
+        '''Return False if controller is to be kept. True if controller
+        was terminated and can be removed from the list.'''
+        controller, keep_connection, on_controller_removed_cb, tid = args
+        if controller.all_ops_completed():
+            logging.debug(
+                'CtrlTerminator._keep_or_terminate()- %s | %s: Disconnecting controller',
+                tid,
+                controller.device,
+            )
+            controller.disconnect(on_controller_removed_cb, keep_connection)
+            return True
+
+        return False
+
+    def _disposal_check(self):
+        # Iterate over the list, terminating (disconnecting) those controllers
+        # that have no pending operations, and remove those controllers from the
+        # list (only keep controllers that still have operations pending).
+        self._controllers[:] = filterfalse(self._keep_or_terminate, self._controllers)
+        disposal_complete = len(self._controllers) == 0
+
+        if disposal_complete:
+            logging.debug('CtrlTerminator._disposal_check()   - Disposal complete')
+            self._audit_tmr.stop()
+            self._udev.unregister_for_action_events('add', self._on_kernel_events)
+            self._udev.unregister_for_action_events('remove', self._on_kernel_events)
+        else:
+            self._audit_tmr.start()  # Restart timer
+
+        return disposal_complete
 
 # ******************************************************************************
 class Service(stas.ServiceABC):
@@ -30,6 +152,7 @@ class Service(stas.ServiceABC):
         sysconf = conf.SysConf()
         self._root = nvme.root()
         self._host = nvme.host(self._root, sysconf.hostnqn, sysconf.hostid, sysconf.hostsymname)
+        self._terminator = CtrlTerminator()
 
         super().__init__(args, default_conf, reload_hdlr)
 
@@ -39,8 +162,31 @@ class Service(stas.ServiceABC):
         logging.debug('Service._release_resources()')
         super()._release_resources()
 
+        if self._terminator:
+            self._terminator.kill()
+
         self._host = None
         self._root = None
+        self._terminator = None
+
+    def _disconnect_all(self):
+        # Tell all controller objects to disconnect
+        keep_connections = self._keep_connections_on_exit()
+        controllers = self._controllers.values()
+        logging.debug(
+            'Service._stop_hdlr()               - Controller count = %s, keep_connections = %s',
+            len(controllers),
+            keep_connections,
+        )
+        for controller in controllers:
+            self._terminator.dispose(controller, self._on_final_disconnect, keep_connections)
+
+    def info(self) -> dict:
+        '''@brief Get the status info for this object (used for debug)'''
+        info = super().info()
+        if self._terminator:
+            info['terminator'] = str(self._terminator.info())
+        return info
 
     @stas.ServiceABC.tron.setter
     def tron(self, value):
@@ -268,7 +414,7 @@ class Stac(Service):
         if not self._alive():
             return GLib.SOURCE_REMOVE
 
-        systemd.daemon.notify('RELOADING=1')
+        sd_notify('RELOADING=1')
         service_cnf = conf.SvcConf()
         service_cnf.reload()
         self.tron = service_cnf.tron
@@ -279,7 +425,7 @@ class Stac(Service):
         for controller in self._controllers.values():
             controller.reload_hdlr()
 
-        systemd.daemon.notify('READY=1')
+        sd_notify('READY=1')
         return GLib.SOURCE_CONTINUE
 
     def _get_log_pages_from_stafd(self):
@@ -507,7 +653,7 @@ class Staf(Service):
         if not self._alive():
             return GLib.SOURCE_REMOVE
 
-        systemd.daemon.notify('RELOADING=1')
+        sd_notify('RELOADING=1')
         service_cnf = conf.SvcConf()
         service_cnf.reload()
         self.tron = service_cnf.tron
@@ -518,7 +664,7 @@ class Staf(Service):
         for controller in self._controllers.values():
             controller.reload_hdlr()
 
-        systemd.daemon.notify('READY=1')
+        sd_notify('READY=1')
         return GLib.SOURCE_CONTINUE
 
     def is_avahi_reported(self, tid):
@@ -620,11 +766,11 @@ class Staf(Service):
         # handled differently and only if the connection cannot be established
         # for a long period of time.
         logging.debug('Staf._config_ctrls_finish()        - must_remove_list     = %s', list(must_remove_list))
-        controllers_to_del = [
+        controllers_to_del = {
             tid
             for tid in controllers_to_del
             if tid in must_remove_list or self._controllers[tid].origin != 'discovered'
-        ]
+        }
 
         logging.debug('Staf._config_ctrls_finish()        - controllers_to_add   = %s', list(controllers_to_add))
         logging.debug('Staf._config_ctrls_finish()        - controllers_to_del   = %s', list(controllers_to_del))
